pyproject.toml
[tool.poetry]
name = "picam"
version = "0.1.0"
description = ""
authors = ["Roland Siegbert <roland.siegbert@phys.ethz.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
django = "^5.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
